Add files via upload
E-Commerce Dashboard

The E-Commerce Dashboard was used to prepare a table of Sales and Profit month-wise in a working sheet. Additionally, a sales table was created region-wise in the same working sheet. A User Control Combo box for Product Category was then implemented to facilitate easy selection. Subsequently, a Column Chart was generated based on the month-wise and region-wise tables. The tables were linked with the combo box for seamless data integration. Finally, a comprehensive dashboard was created, showcasing the sales, profit, and product category information in a visually appealing manner.